pyproject.toml
[project]
name = "luminary"
version = "0.1.0"
description = "AI Code Reviewer for GitLab"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Luminary Team"}
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-gitlab>=4.0.0",
    "requests>=2.31.0",
    "tenacity>=8.2.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
luminary = "luminary.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
